refactor to dotenv/ssm class
