+++ lib/jets/dotenv.rb
@@ -1,9 +1,6 @@
 require 'dotenv'
-require 'aws-sdk-ssm'
 
 class Jets::Dotenv
-  SSM_VARIABLE_REGEXP = /^ssm:(.*)/
-
   def self.load!(remote=false)
     new(remote).load!
   end
@@ -14,8 +11,8 @@ class Jets::Dotenv
   end
 
   def load!
-    env = ::Dotenv.load(*dotenv_files)
-    interpolate_ssm_variables(env)
+    vars = ::Dotenv.load(*dotenv_files)
+    Ssm.new(vars).interpolate!
   end
 
   # dotenv files with the following precedence:
@@ -39,35 +36,5 @@ class Jets::Dotenv
 
   def root
     Jets.root || Pathname.new(ENV["JETS_ROOT"] || Dir.pwd)
-  end
-
-  private
-
-  def interpolate_ssm_variables(variables)
-    interpolated_variables = variables.map do |key, value|
-      if value[SSM_VARIABLE_REGEXP]
-        value = fetch_ssm_value($1)
-      end
-
-      [key, value]
-    end
-
-    interpolated_variables.each do |key, value|
-      ENV[key] = value
-    end
-
-    interpolated_variables.to_h
-  end
-
-  def fetch_ssm_value(name)
-    name = "/#{Jets.application.config.project_name}/#{Jets.env}/#{name}" unless name.start_with?("/")
-    response = ssm.get_parameter(name: name, with_decryption: true)
-    response.parameter.value
-  rescue Aws::SSM::Errors::ParameterNotFound
-    abort "Error loading .env variables. No parameter matching #{name} found on AWS SSM.".color(:red)
-  end
-
-  def ssm
-    @ssm ||= Aws::SSM::Client.new
   end
 end
